+++ isaw/awol/normalize_space.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+Normalize space in a string.
+"""
+
+import re
+
+def normalize_space(raw):
+    """Flatten all whitespace in a string."""
+
+    rx = re.compile('\s+')
+    cooked = rx.sub(' ', raw).strip()
+    return cooked
+
+
